changing some db types

+++ internal/infra/database/postgres/migrations/000001_create-parts-table.sql
@@ -6,12 +6,12 @@ CREATE TABLE parts (
     category VARCHAR(50) NOT NULL,
     current_stock INTEGER NOT NULL DEFAULT 0,
     minimum_stock INTEGER NOT NULL DEFAULT 0,
-    average_daily_sales NUMERIC(10,2) NOT NULL DEFAULT 0,
+    average_daily_sales INTEGER NOT NULL DEFAULT 0,
     lead_time_days INTEGER NOT NULL DEFAULT 0,
     unit_cost NUMERIC(12,2) NOT NULL DEFAULT 0,
     criticality_level INTEGER NOT NULL DEFAULT 1,
     projected_stock INTEGER NOT NULL DEFAULT 0,
-    urgency_score NUMERIC(10,2) NOT NULL DEFAULT 0,
+    urgency_score INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
     updated_at TIMESTAMP WITH TIME ZONE DEFAULT NOW()
 );
